requirements.txt
torch==2.6.0+cu126 --index-url https://download.pytorch.org/whl/cu126
torchvision --index-url https://download.pytorch.org/whl/cu126
torchaudio --index-url https://download.pytorch.org/whl/cu126

ipython==8.12.3
matplotlib==3.8.2
pandas==2.0.3
pillow >= 9.5.0
plotly==6.0.0
Requests==2.32.3
safetensors==0.5.3
scikit_learn==1.3.0
simple-parsing >= 0.1.5
ultralytics==8.3.85
diffusers
huggingface-hub
numpy
simple-parsing >= 0.1.5
timm @ git+https://github.com/huggingface/pytorch-image-models@main#egg=timm
tokenizers
torch >= 2.0.0
torchvision
transformers
nbformat>=4.2.0